requirements.txt
fastapi[standard]
pyjwt
SQLAlchemy
cachetools